requirements.txt
opencv-python
sounddevice
customtkinter
Pillow
numpy
